add ToDo server
rename main.js => todo.js; create new main.js with serverpart
