+++ miamoremustachio/ToDo/modules/constants.js
@@ -1,3 +1,5 @@
+const PORT = 3000;
+
 const STATUSES = {
   TO_DO: "to do",
   IN_PROGRESS: "in progress",
@@ -19,24 +21,26 @@ const { LOW, HIGH } = PRIORITIES;
 const { MIN: MIN_LENGTH, MAX: MAX_LENGTH } = TITLE_LENGTH;
 
 const ERROR_MESSAGES = {
-  INVALID_TITLE: "Invalid title (only strings are allowed).",
-  INVALID_STATUS: `Invalid status (use "${TO_DO}", "${IN_PROGRESS}" or "${DONE}").`,
-  INVALID_PRIORITY: `Invalid priority (use "${LOW}" or "${HIGH}")`,
-  INVALID_NUMBER: "Invalid task number (only finite numbers are allowed)",
-  INCORRECT_TITLE_LENGTH: `Incorrect task's length (only titles between ${MIN_LENGTH} and ${MAX_LENGTH} characters are allowed).`,
-  MISSING_ARGUMENTS: "Arguments is missing.",
-  TASK_EXISTS: "Task you want to add is already in list.",
-  TASK_NOT_FOUND: "Task wasn't found in list.",
+  INVALID_TITLE: "✗ Invalid title (only strings are allowed).",
+  INVALID_STATUS: `✗ Invalid status (use "${TO_DO}", "${IN_PROGRESS}" or "${DONE}").`,
+  INVALID_PRIORITY: `✗ Invalid priority (use "${LOW}" or "${HIGH}")`,
+  INVALID_NUMBER: "✗ Invalid task number (only finite numbers are allowed)",
+  INCORRECT_TITLE_LENGTH: `✗ Incorrect task's length (only titles between ${MIN_LENGTH} and ${MAX_LENGTH} characters are allowed).`,
+  MISSING_ARGUMENTS: "✗ Arguments is missing.",
+  TASK_EXISTS: "✗ Task you want to add is already in list.",
+  TASK_NOT_FOUND: "✗ Task wasn't found in list.",
 };
 
 const INFO_MESSAGES = {
-  SUCCESSFULLY_ADDED: "Task has successfully added:",
-  SUCCESSFULLY_CHANGED_STATUS: "Task status has successfully changed:",
-  SUCCESSFULLY_CHANGED_PRIORITY: "Task priority has successfully changed:",
-  SUCCESSFULLY_DELETED: "Task has successfully deleted.",
+  START_MESSAGE: "ToDo server is running.",
+  SUCCESSFULLY_ADDED: "✓ Task has successfully added.",
+  SUCCESSFULLY_CHANGED_STATUS: "✓ Task status has successfully changed.",
+  SUCCESSFULLY_CHANGED_PRIORITY: "✓ Task priority has successfully changed.",
+  SUCCESSFULLY_DELETED: "✓ Task has successfully deleted.",
 };
 
 module.exports = {
+  PORT,
   STATUSES,
   PRIORITIES,
   TITLE_LENGTH,
