+++ miamoremustachio/ToDo/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "todo_server",
+  "version": "1.0.0",
+  "description": "Server version of ToDo app written in Express.js",
+  "main": "main.js",
+  "scripts": {
+    "test": "node main.js"
+  },
+  "author": "miamoremustachio",
+  "license": "ISC",
+  "dependencies": {
+    "express": "^4.18.2"
+  }
+}
